+++ src/components/TakeHomeCalculator/TakeHomeChart.tsx
@@ -142,6 +142,7 @@ interface TakeHomeChartProps {
   isEmploymentIncome: boolean;
   isSubjectToLongTermCarePremium: boolean;
   healthInsuranceProvider: HealthInsuranceProviderId;
+  nationalPensionExemption?: boolean;
   region: string;
   dcPlanContributions: number;
   dependents: Dependent[];
@@ -221,6 +222,7 @@ const TakeHomeChart: React.FC<TakeHomeChartProps> = ({
   isEmploymentIncome,
   isSubjectToLongTermCarePremium,
   healthInsuranceProvider,
+  nationalPensionExemption,
   region,
   dcPlanContributions,
   dependents,
@@ -319,6 +321,7 @@ const TakeHomeChart: React.FC<TakeHomeChartProps> = ({
         incomeYear,
         isSubjectToLongTermCarePremium,
         healthInsuranceProvider,
+        nationalPensionExemption,
         region,
         dcPlanContributions,
         dependents,
@@ -337,6 +340,7 @@ const TakeHomeChart: React.FC<TakeHomeChartProps> = ({
       incomeYear,
       isSubjectToLongTermCarePremium,
       healthInsuranceProvider,
+      nationalPensionExemption,
       region,
       dcPlanContributions,
       dependents,
@@ -392,6 +396,7 @@ const TakeHomeChart: React.FC<TakeHomeChartProps> = ({
                   isEmploymentIncome,
                   isSubjectToLongTermCarePremium,
                   healthInsuranceProvider,
+                  nationalPensionExemption,
                   region,
                   dcPlanContributions,
                   dependents,
@@ -433,6 +438,7 @@ const TakeHomeChart: React.FC<TakeHomeChartProps> = ({
     isEmploymentIncome,
     isSubjectToLongTermCarePremium,
     healthInsuranceProvider,
+    nationalPensionExemption,
     region,
     dcPlanContributions,
     dependents,